Add override for Limbu U+193A
From: Martin Hosken 
Sent: 25 August 2019 19:35
In Limbu the length mark (U+193A) currently has an Indic Syllabic Category of Vowel_Dependent which is wrong. It is a vowel modifier (for which there is no Indic Syllabic Category). This means that an error circle is inserted for correctly ordered text. We are in the process of writing the proposal to change the category of the character.